fix: performance improvements
