+++ web/app/(scoped)/orgs/[orgSlug]/workspaces/[workspaceSlug]/dashboards/dashboard-view.tsx
@@ -2,6 +2,7 @@
 
 import * as React from "react"
 import type { Route } from "next"
+import dynamic from "next/dynamic"
 import Link from "next/link"
 import { usePathname, useRouter, useSearchParams } from "next/navigation"
 import { keepPreviousData, queryOptions, useQuery } from "@tanstack/react-query"
@@ -12,27 +13,6 @@ import {
   useReactTable,
 } from "@tanstack/react-table"
 import { BotIcon, LayoutDashboard, RefreshCw } from "lucide-react"
-import {
-  Area,
-  AreaChart,
-  Bar,
-  BarChart,
-  CartesianGrid,
-  Cell,
-  Label,
-  Legend,
-  Line,
-  LineChart,
-  Pie,
-  PieChart,
-  PolarAngleAxis,
-  RadialBar,
-  RadialBarChart,
-  Scatter,
-  ScatterChart,
-  XAxis,
-  YAxis,
-} from "recharts"
 import {
   listDashboardTableRows,
   queryDashboard,
@@ -47,7 +27,6 @@ import { AdminDataGrid, type AdminColumnLayout } from "@/components/admin-data-g
 import { TablePagination } from "@/components/table-pagination"
 import { Button } from "@/components/ui/button"
 import { DateRangePicker } from "@/components/ui/calendar"
-import { ChartContainer, ChartTooltip, type ChartConfig } from "@/components/ui/chart"
 import {
   Select,
   SelectContent,
@@ -56,18 +35,19 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select"
+import { Skeleton } from "@/components/ui/skeleton"
 import { Spinner } from "@/components/ui/spinner"
 import { EmptyValue, RelativeDateTime } from "@/components/ui/table"
 import { dayjs } from "@/lib/format"
 import { cn } from "@/lib/utils"
 
-const palette = [
-  "var(--chart-1)",
-  "var(--chart-2)",
-  "var(--chart-3)",
-  "var(--chart-4)",
-  "var(--chart-5)",
-] as const
+const DashboardChart = dynamic(
+  () => import("./dashboard-chart").then((module) => module.DashboardChart),
+  {
+    loading: () => <Skeleton className="h-full w-full" />,
+    ssr: false,
+  }
+)
 
 export function DashboardView({
   dashboard,
@@ -231,6 +211,29 @@ function Widget({
   workspaceId: string
   workspacePath: string
 }) {
+  const [seen, setSeen] = React.useState(false)
+  const sectionRef = React.useRef<HTMLElement>(null)
+
+  React.useEffect(() => {
+    const section = sectionRef.current
+    if (!section) return
+
+    const observer = new IntersectionObserver(
+      (entries) => {
+        for (const entry of entries) {
+          if (!entry.isIntersecting) continue
+          // Keep charts mounted after first paint so revisiting them does not
+          // move Recharts initialization work back onto the scroll path.
+          setSeen(true)
+          observer.disconnect()
+        }
+      },
+      { rootMargin: "256px 0px" }
+    )
+    observer.observe(section)
+    return () => observer.disconnect()
+  }, [])
+
   const width =
     widget.width === "full"
       ? "col-span-12"
@@ -238,7 +241,13 @@ function Widget({
         ? "col-span-12 lg:col-span-6"
         : "col-span-12 md:col-span-6 xl:col-span-4"
   return (
-    <section className={cn("bg-background h-80 min-w-0 overflow-hidden rounded-sm border", width)}>
+    <section
+      className={cn(
+        "bg-background h-80 min-w-0 overflow-hidden rounded-sm border [content-visibility:auto]",
+        width
+      )}
+      ref={sectionRef}
+    >
       <header className="flex h-11 items-center border-b px-4">
         <h2 className="truncate text-sm font-semibold">{widget.title}</h2>
       </header>
@@ -263,8 +272,10 @@ function Widget({
           <div className="text-muted-foreground flex h-full items-center justify-center text-sm">
             No data
           </div>
+        ) : seen ? (
+          <DashboardChart widget={widget} result={result} />
         ) : (
-          <Chart widget={widget} result={result} />
+          <Skeleton className="h-full w-full" />
         )}
       </div>
     </section>
@@ -292,180 +303,6 @@ function InvalidWidget({
         </Link>
       </Button>
     </div>
-  )
-}
-
-function Chart({
-  widget,
-  result,
-}: {
-  widget: DashboardWidget
-  result: DashboardWidgetQueryResult
-}) {
-  const config = Object.fromEntries(
-    widget.series.map((series, index) => [
-      `s${index}`,
-      { label: series.label, color: palette[index] },
-    ])
-  ) satisfies ChartConfig
-  if (widget.kind === "line" || widget.kind === "step" || widget.kind === "area") {
-    const data = result.points.map((point) =>
-      Object.assign(
-        { at: dayjs(point.at).format("lll") },
-        Object.fromEntries(point.values.map((value, index) => [`s${index}`, value]))
-      )
-    )
-    if (widget.kind === "area") {
-      return (
-        <ChartContainer className="h-full w-full" config={config}>
-          <AreaChart data={data}>
-            <CartesianGrid vertical={false} />
-            <XAxis dataKey="at" minTickGap={28} />
-            <YAxis width={40} />
-            <ChartTooltip />
-            <Legend />
-            {widget.series.map((series, index) => (
-              <Area
-                key={series.name}
-                dataKey={`s${index}`}
-                fill={`var(--color-s${index})`}
-                fillOpacity={0.14}
-                name={series.label}
-                stroke={`var(--color-s${index})`}
-                type="monotone"
-              />
-            ))}
-          </AreaChart>
-        </ChartContainer>
-      )
-    }
-    return (
-      <ChartContainer className="h-full w-full" config={config}>
-        <LineChart data={data}>
-          <CartesianGrid vertical={false} />
-          <XAxis dataKey="at" minTickGap={28} />
-          <YAxis width={40} />
-          <ChartTooltip />
-          <Legend />
-          {widget.series.map((series, index) => (
-            <Line
-              key={series.name}
-              dataKey={`s${index}`}
-              dot={false}
-              name={series.label}
-              stroke={`var(--color-s${index})`}
-              strokeWidth={2}
-              type={widget.kind === "step" ? "stepAfter" : "monotone"}
-            />
-          ))}
-        </LineChart>
-      </ChartContainer>
-    )
-  }
-  if (widget.kind === "pie") {
-    const data = result.categories.map((category) => ({
-      name: category.label,
-      value: category.values[0],
-    }))
-    return (
-      <ChartContainer className="h-full w-full" config={config}>
-        <PieChart>
-          <ChartTooltip />
-          <Legend />
-          <Pie data={data} dataKey="value" nameKey="name" innerRadius={52} outerRadius={92}>
-            {data.map((item, index) => (
-              <Cell key={item.name} fill={palette[index % palette.length]} />
-            ))}
-          </Pie>
-        </PieChart>
-      </ChartContainer>
-    )
-  }
-  if (widget.kind === "bar" || widget.kind === "horizontal_grouped_bar") {
-    const data = result.categories.map((category) =>
-      Object.assign(
-        { category: category.label },
-        Object.fromEntries(category.values.map((value, index) => [`s${index}`, value]))
-      )
-    )
-    const horizontal = widget.kind === "horizontal_grouped_bar"
-    return (
-      <ChartContainer className="h-full w-full" config={config}>
-        <BarChart data={data} layout={horizontal ? "vertical" : "horizontal"}>
-          <CartesianGrid horizontal={!horizontal} vertical={horizontal} />
-          {horizontal ? (
-            <>
-              <XAxis type="number" />
-              <YAxis dataKey="category" type="category" width={86} />
-            </>
-          ) : (
-            <>
-              <XAxis dataKey="category" />
-              <YAxis width={40} />
-            </>
-          )}
-          <ChartTooltip />
-          <Legend />
-          {widget.series.map((series, index) => (
-            <Bar
-              key={series.name}
-              dataKey={`s${index}`}
-              fill={`var(--color-s${index})`}
-              name={series.label}
-              radius={2}
-            />
-          ))}
-        </BarChart>
-      </ChartContainer>
-    )
-  }
-  if (widget.kind === "scatter") {
-    return (
-      <ChartContainer className="h-full w-full" config={config}>
-        <ScatterChart>
-          <CartesianGrid />
-          <XAxis dataKey="x" type="number" />
-          <YAxis dataKey="y" type="number" width={40} />
-          <ChartTooltip />
-          <Legend />
-          {widget.series.map((series, index) => (
-            <Scatter
-              key={series.name}
-              data={result.scatter.filter((point) => point.series === index)}
-              fill={`var(--color-s${index})`}
-              name={series.label}
-            />
-          ))}
-        </ScatterChart>
-      </ChartContainer>
-    )
-  }
-  const minimum = widget.minimum ?? 0
-  const maximum = widget.maximum ?? 100
-  const value = result.value ?? minimum
-  const percent = Math.max(0, Math.min(100, ((value - minimum) / (maximum - minimum)) * 100))
-  const tone = widget.thresholds.findLast((threshold) => value >= threshold.value)?.tone
-  const fill =
-    tone === "critical" ? "var(--destructive)" : tone === "warning" ? "var(--warning)" : palette[0]
-  return (
-    <ChartContainer className="h-full w-full" config={config}>
-      <RadialBarChart
-        data={[{ name: widget.series[0]?.label ?? widget.title, value: percent, fill }]}
-        innerRadius="72%"
-        outerRadius="100%"
-        startAngle={210}
-        endAngle={-30}
-      >
-        <PolarAngleAxis angleAxisId={0} domain={[0, 100]} tick={false} type="number" />
-        <RadialBar dataKey="value" background cornerRadius={4} />
-        <Label
-          className="fill-foreground text-3xl font-semibold tabular-nums"
-          position="center"
-          value={value.toLocaleString()}
-        />
-        <Legend />
-      </RadialBarChart>
-    </ChartContainer>
   )
 }
